requirements_GP1.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
addict=2.4.0=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
ansi2html=1.8.0=pypi_0
anyio=3.5.0=py310h06a4308_0
appdirs=1.4.4=pypi_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h7f8727e_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=22.1.0=py310h06a4308_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=py310h06a4308_0
bleach=4.1.0=pyhd3eb1b0_0
bytesbufio=1.0.3=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2022.12.7=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=3.1.0=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
cmake=3.26.3=pypi_0
colorama=0.4.6=pypi_0
comm=0.1.4=pypi_0
configargparse=1.7=pypi_0
dacite=1.8.1=pypi_0
dash=2.13.0=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
deprecated=1.2.14=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
efficientnet-pytorch=0.7.1=pypi_0
elasticdeform=0.5.0=pypi_0
entrypoints=0.4=py310h06a4308_0
et-xmlfile=1.1.0=pypi_0
executing=1.2.0=pypi_0
filelock=3.12.0=pypi_0
flask=2.2.5=pypi_0
flatbuffers=23.5.26=pypi_0
flit-core=3.8.0=py310h06a4308_0
frozenlist=1.4.1=pypi_0
fsspec=2023.5.0=pypi_0
fuzzywuzzy=0.18.0=pypi_0
gast=0.4.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.42=pypi_0
google-auth=2.21.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
greenlet=2.0.2=pypi_0
grpcio=1.56.0=pypi_0
h5py=3.10.0=pypi_0
htmlmin=0.1.12=pypi_0
huggingface-hub=0.15.1=pypi_0
humanize=4.8.0=pypi_0
icu=58.2=he6710b0_3
idna=3.4=py310h06a4308_0
imagehash=4.3.1=pypi_0
imageio=2.31.0=pypi_0
inputimeout=1.0.4=pypi_0
ipykernel=6.19.2=py310h2f386ee_0
ipython=8.12.0=py310h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.1.0=pypi_0
itk=5.3.0=pypi_0
itk-core=5.3.0=pypi_0
itk-filtering=5.3.0=pypi_0
itk-io=5.3.0=pypi_0
itk-numerics=5.3.0=pypi_0
itk-registration=5.3.0=pypi_0
itk-segmentation=5.3.0=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.18.1=py310h06a4308_1
jinja2=3.1.2=py310h06a4308_0
joblib=1.2.0=pypi_0
jsonschema=4.17.3=py310h06a4308_0
jupyter=1.0.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter_client=8.1.0=py310h06a4308_0
jupyter_core=5.3.0=py310h06a4308_0
jupyter_server=1.23.4=py310h06a4308_0
jupyterlab-widgets=3.0.8=pypi_0
jupyterlab_pygments=0.1.2=py_0
keras=2.13.1=pypi_0
kneed=0.8.5=pypi_0
lazy-loader=0.2=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=16.0.0=pypi_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxml2=2.10.3=hcbfbd50_0
libxslt=1.1.37=h2085143_0
lightning-utilities=0.11.2=pypi_0
lit=16.0.5=pypi_0
llvmlite=0.41.1=pypi_0
lxml=4.9.2=py310h5eee18b_0
markdown=3.4.3=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.1=py310h7f8727e_0
matplotlib-inline=0.1.6=py310h06a4308_0
mdurl=0.1.2=pypi_0
mistune=0.8.4=py310h7f8727e_1000
monai=1.3.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
multimethod=1.9.1=pypi_0
munch=3.0.0=pypi_0
nbclassic=0.5.4=py310h06a4308_0
nbclient=0.5.13=py310h06a4308_0
nbconvert=6.5.4=py310h06a4308_0
nbformat=5.7.0=py310h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py310h06a4308_0
networkx=3.1=pypi_0
nibabel=5.1.0=pypi_0
notebook=6.5.3=py310h06a4308_0
notebook-shim=0.2.2=py310h06a4308_0
numba=0.58.1=pypi_0
numpy=1.25.2=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
oauthlib=3.2.2=pypi_0
open3d=0.17.0=pypi_0
opencv-python=4.7.0.72=pypi_0
openpyxl=3.1.2=pypi_0
openssl=1.1.1t=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=23.0=py310h06a4308_0
pandas=2.0.1=pypi_0
pandas-profiling=3.6.6=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
patsy=0.5.3=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
phik=0.12.3=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pip=23.3=pypi_0
platformdirs=2.5.2=py310h06a4308_0
plotly=5.16.1=pypi_0
pretrainedmodels=0.7.4=pypi_0
pretty-errors=1.2.25=pypi_0
prometheus_client=0.14.1=py310h06a4308_0
prompt-toolkit=3.0.36=py310h06a4308_0
protobuf=4.23.4=pypi_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.12=pypi_0
pydicom=2.4.2=pypi_0
pygments=2.16.1=pypi_0
pykwalify=1.8.0=pypi_0
pylidc=0.2.3=pypi_0
pyperclip=1.8.2=pypi_0
pyquaternion=0.9.9=pypi_0
pyradiomics=3.0.1=pypi_0
pyrsistent=0.18.0=py310h7f8727e_0
python=3.10.10=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py310h06a4308_0
python-graphviz=0.20.1=pypi_0
pytimedinput=2.0.1=pypi_0
pytorch-lightning=2.2.4=pypi_0
pytz=2023.3=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py310h6a678d5_0
qtconsole=5.4.4=pypi_0
qtpy=2.4.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
retrying=1.3.4=pypi_0
rich=13.6.0=pypi_0
rsa=4.9=pypi_0
ruamel-yaml=0.17.32=pypi_0
ruamel-yaml-clib=0.2.7=pypi_0
safetensors=0.3.1=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.11.1=pypi_0
seaborn=0.12.2=pypi_0
segmentation-models-pytorch=0.3.3=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
sentry-sdk=1.42.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=65.6.3=py310h06a4308_0
shap=0.43.0=pypi_0
shellingham=1.5.4=pypi_0
simpleitk=2.3.1=pypi_0
six=1.16.0=pyhd3eb1b0_1
slicer=0.0.7=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.2.0=py310h06a4308_1
soupsieve=2.4=py310h06a4308_0
sqlalchemy=2.0.19=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
statsmodels=0.14.0=pypi_0
sympy=1.12=pypi_0
tangled-up-in-unicode=0.2.0=pypi_0
tenacity=8.2.3=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorflow-estimator=2.13.0=pypi_0
tensorflow-io-gcs-filesystem=0.32.0=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=py310h06a4308_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.4.12=pypi_0
timm=0.9.2=pypi_0
tinycss2=1.2.1=py310h06a4308_0
tk=8.6.12=h1ccaba5_0
torch=2.0.1=pypi_0
torchaudio=2.0.2=pypi_0
torchio=0.19.1=pypi_0
torchmetrics=1.4.0=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.15.2=pypi_0
torchviz=0.0.2=pypi_0
tornado=6.2=py310h5eee18b_0
tqdm=4.65.0=pypi_0
traitlets=5.7.1=py310h06a4308_0
triton=2.0.0=pypi_0
typeguard=2.13.3=pypi_0
typer=0.9.0=pypi_0
typing-extensions=4.4.0=py310h06a4308_0
typing_extensions=4.4.0=py310h06a4308_0
tzdata=2023.3=pypi_0
urllib3=1.26.16=pypi_0
visions=0.7.5=pypi_0
wandb=0.16.4=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310h06a4308_1
websocket-client=0.58.0=py310h06a4308_4
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py310h06a4308_0
widgetsnbextension=4.0.8=pypi_0
wordcloud=1.9.2=pypi_0
wrapt=1.15.0=pypi_0
xz=5.2.10=h5eee18b_1
yarl=1.9.4=pypi_0
ydata-profiling=4.5.1=pypi_0
zeromq=4.3.4=h2531618_0
zlib=1.2.13=h5eee18b_0
